+++ time_stamp.h
@@ -34,5 +34,5 @@ typedef struct
 
 
 void free_time_management(TimeManagement *all_time);
-TimeManagement *init_time_management(int size);
+TimeManagement *init_time_management(size_t size);
 #endif
